fix(util): split utility functions for more clarity
