+++ internal/resource/utils/secret.go
